Reduce code duplication in simulator pages
AbstractSimulatorPage base class introduced
